Separando as classes para Simplificar a Main
