+++ src/main/java/com/audiomania/estoque/Main.java
@@ -1,4 +1,8 @@
 package com.audiomania.estoque;
 
 public class Main {
+    public static void main(String[] args) {
+        GerenciadorProdutos gerenciador = new GerenciadorProdutos();
+        gerenciador.exibirMenu();
+    }
 }
